ai_info_assistant/pyproject.toml
[project]
name = "ai-info-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.128.0",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
    "rich>=13.0.0",
    "openai>=2.15.0",
    "google-genai>=1.56.0",
    "anthropic>=0.75.0",
    "portkey-ai>=2.1.0",
    "requests>=2.32.5",
]

[tool.uv.workspace]
members = [
    "market_intel_ai",
]
